XCode compiler warnings fix

+++ Classes/ShareKit/Core/Helpers/OAuth/OAMutableURLRequest.m
@@ -192,7 +192,7 @@ signatureProvider:(id<OASignatureProviding, NSObject>)aProvider
 
 - (void)_generateTimestamp
 {
-    timestamp = [[NSString stringWithFormat:@"%d", time(NULL)] retain];
+    timestamp = [[NSString stringWithFormat:@"%d", (NSInteger)time(NULL)] retain];
 }
 
 - (void)_generateNonce
